Fix for Mac build of note example.

+++ src/program/example/note/playback.cpp
@@ -173,7 +173,7 @@ void
 GltPlayback::copy(void *buffer,uint32 size)
 {
     memset(buffer,0,size);
-    memcpy(buffer,&_buffer[0],std::min(_buffer.size(),size-6));
+    memcpy(buffer,&_buffer[0],std::min(_buffer.size(),size_t(size)-6));
 }
